Reworked statistics without authorization Added validation for server configuration
